Use a idle handler for loading songs inside folders
We iterate over the list of files and then lazily load each song and
queue it up using an idle handler in the main loop.

When loading songs we are I/O limited, so using a separate thread does
not buy us anything.

We are still traversing the folder in the main loop, so the next step
would be to iterate over the contents of a folder in a future, and at
the end lazily load the songs into the queue.